+++ src/main/java/allclear/dto/requestDto/subject/GetSubjectListRequestDto.java
@@ -14,7 +14,7 @@ public class GetSubjectListRequestDto {
     private String professor; //교수명
     private String department; //개설 학과
     private String majorClassification; //이수 구분(주전공)
-    private Integer credit; //학점
+    private Double credit; //학점
     private String subjectTarget; //수강 대상
 
     /**시간에 따른 조회 기능 추가 예정*/
